Split the Gradio UI into a training page and a dataset annotation page.
