+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "accelerate~=0.25.0",
     "datasets~=2.14.3",
     "diffusers~=0.25.0",
+    "fastapi",
     "gradio",
     "numpy",
     "omegaconf",
@@ -32,6 +33,7 @@ dependencies = [
     "torchvision",
     "tqdm",
     "transformers~=4.36.0",
+    "uvicorn[standard]",
     "xformers>=0.0.23",
 ]
