+++ src/invoke_training/scripts/invoke_train_ui.py
@@ -1,9 +1,11 @@
-from invoke_training.ui.app import App
+import uvicorn
+
+from invoke_training.ui.app import build_app
 
 
 def main():
-    app = App()
-    app.launch()
+    app = build_app()
+    uvicorn.run(app)
 
 
 if __name__ == "__main__":
