+++ src/invoke_training/ui/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>invoke-training</title>
+</head>
+<body>
+    <h1>invoke-training</h1>
+    <p>
+        <a href="/train">Training</a>
+    </p>
+    <p>
+        <a href="/data">Dataset Annotation</a>
+    </p>
+</body>
+</html>
